+++ src/memory_seam/agent_packages/memory_librarian/__init__.py
@@ -1,0 +1,5 @@
+"""Memory librarian template package/starter resources."""
+
+PACKAGE_NAME = "memory_librarian"
+
+__all__ = ["PACKAGE_NAME"]